+++ changelog.md
@@ -2,6 +2,7 @@ v0.4.1-beta
 - Changed getEnabled to isEnabled inside UIClickable.
 - Fixed issues related to uiScale.
 - Added uiScale as a constant.
+- Modified DockPanel to have an inner panel.
 
 v0.4.0-beta
 - package changes.